Refactor `Deployment` func for different platforms
This commit refactors how k8s Deployments are provisioned depending
on the k8s "distribution" (vanilla or ocp). It creates an interface
as an abstraction of the different deployments and a struct (DeploymentObj)
that references the concrete types.
ref #884
[noissue]
