+++ controllers/repo_manager/api.go
@@ -19,10 +19,12 @@ package repo_manager
 import (
 	"context"
 	"fmt"
-	"os"
 	"strconv"
 	"time"
 
+	"github.com/go-logr/logr"
+	repomanagerpulpprojectorgv1beta2 "github.com/pulp/pulp-operator/apis/repo-manager.pulpproject.org/v1beta2"
+	"github.com/pulp/pulp-operator/controllers"
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
 	appsv1 "k8s.io/api/apps/v1"
@@ -32,10 +34,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
-
-	"github.com/go-logr/logr"
-	repomanagerpulpprojectorgv1beta2 "github.com/pulp/pulp-operator/apis/repo-manager.pulpproject.org/v1beta2"
-	"github.com/pulp/pulp-operator/controllers"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -95,6 +93,9 @@ func (r *RepoManagerReconciler) pulpApiController(ctx context.Context, pulp *rep
 		pulp.Spec.ContainerTokenSecret = pulp.Name + "-container-auth"
 		r.Patch(ctx, pulp, patch)
 	}
+
+	// define the k8s Deployment function based on k8s distribution and deployment type
+	deploymentForPulpApi := initDeployment(API_DEPLOYMENT).deploy
 
 	// list of pulp-api resources that should be provisioned
 	resources := []ApiResource{
@@ -201,599 +202,6 @@ func fileStoragePVC(resources FunctionResources) client.Object {
 	// Set Pulp instance as the owner and controller
 	ctrl.SetControllerReference(resources.Pulp, pvc, resources.RepoManagerReconciler.Scheme)
 	return pvc
-}
-
-// deploymentForPulpApi returns a pulp-api Deployment object
-func deploymentForPulpApi(resources FunctionResources) client.Object {
-	replicas := resources.Pulp.Spec.Api.Replicas
-	ls := labelsForPulpApi(resources.Pulp)
-
-	affinity := &corev1.Affinity{}
-	if resources.Pulp.Spec.Api.Affinity != nil {
-		affinity = resources.Pulp.Spec.Api.Affinity
-	}
-
-	if resources.Pulp.Spec.Affinity != nil { // [DEPRECATED] Temporarily adding to keep compatibility with ansible version.
-		affinity.NodeAffinity = resources.Pulp.Spec.Affinity.NodeAffinity
-	}
-
-	// if no strategy is defined in pulp CR we are setting `strategy.Type` with the
-	// default value ("RollingUpdate"), this will be helpful during the reconciliation
-	// when a strategy was previously defined and eventually the field is removed
-	strategy := resources.Pulp.Spec.Api.Strategy
-	if strategy.Type == "" {
-		strategy.Type = "RollingUpdate"
-	}
-
-	// pulp image is built to run with user 0
-	// we are enforcing the containers to run as 1000
-	runAsUser := int64(700)
-	fsGroup := int64(700)
-	podSecurityContext := &corev1.PodSecurityContext{}
-	IsOpenShift, _ := controllers.IsOpenShift()
-	if !IsOpenShift {
-		podSecurityContext = &corev1.PodSecurityContext{
-			RunAsUser: &runAsUser,
-			FSGroup:   &fsGroup,
-		}
-	}
-
-	nodeSelector := map[string]string{}
-	if resources.Pulp.Spec.Api.NodeSelector != nil {
-		nodeSelector = resources.Pulp.Spec.Api.NodeSelector
-	} else if resources.Pulp.Spec.NodeSelector != nil { // [DEPRECATED] Temporarily adding to keep compatibility with ansible version.
-		nodeSelector = resources.Pulp.Spec.NodeSelector
-	}
-
-	toleration := []corev1.Toleration{}
-	if resources.Pulp.Spec.Api.Tolerations != nil {
-		toleration = resources.Pulp.Spec.Api.Tolerations
-	} else if resources.Pulp.Spec.Tolerations != nil { // [DEPRECATED] Temporarily adding to keep compatibility with ansible version.
-		toleration = resources.Pulp.Spec.Tolerations
-	}
-
-	topologySpreadConstraint := []corev1.TopologySpreadConstraint{}
-	if resources.Pulp.Spec.Api.TopologySpreadConstraints != nil {
-		topologySpreadConstraint = resources.Pulp.Spec.Api.TopologySpreadConstraints
-	} else if resources.Pulp.Spec.TopologySpreadConstraints != nil {
-		topologySpreadConstraint = resources.Pulp.Spec.TopologySpreadConstraints
-	}
-
-	gunicornWorkers := strconv.Itoa(resources.Pulp.Spec.Api.GunicornWorkers)
-	if resources.Pulp.Spec.GunicornAPIWorkers > 0 { // [DEPRECATED] Temporarily adding to keep compatibility with ansible version.
-		gunicornWorkers = strconv.Itoa(resources.Pulp.Spec.GunicornAPIWorkers)
-	}
-	gunicornTimeout := strconv.Itoa(resources.Pulp.Spec.Api.GunicornTimeout)
-	if resources.Pulp.Spec.GunicornTimeout > 0 { // [DEPRECATED] Temporarily adding to keep compatibility with ansible version.
-		gunicornTimeout = strconv.Itoa(resources.Pulp.Spec.GunicornTimeout)
-	}
-	envVars := []corev1.EnvVar{
-		{Name: "PULP_GUNICORN_TIMEOUT", Value: gunicornTimeout},
-		{Name: "PULP_API_WORKERS", Value: gunicornWorkers},
-	}
-
-	var dbHost, dbPort string
-
-	// if there is no ExternalDBSecret defined, we should
-	// use the postgres instance provided by the operator
-	if len(resources.Pulp.Spec.PostgresConfigurationSecret) > 0 { // [DEPRECATED] Temporarily adding to keep compatibility with ansible version.
-		postgresEnvVars := []corev1.EnvVar{
-			{
-				Name: "POSTGRES_SERVICE_HOST",
-				ValueFrom: &corev1.EnvVarSource{
-					SecretKeyRef: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: resources.Pulp.Spec.PostgresConfigurationSecret,
-						},
-						Key: "POSTGRES_HOST",
-					},
-				},
-			}, {
-				Name: "POSTGRES_SERVICE_PORT",
-				ValueFrom: &corev1.EnvVarSource{
-					SecretKeyRef: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: resources.Pulp.Spec.PostgresConfigurationSecret,
-						},
-						Key: "POSTGRES_PORT",
-					},
-				},
-			},
-		}
-		envVars = append(envVars, postgresEnvVars...)
-	} else if len(resources.Pulp.Spec.Database.ExternalDBSecret) == 0 {
-		containerPort := 0
-		if resources.Pulp.Spec.Database.PostgresPort == 0 {
-			containerPort = 5432
-		} else {
-			containerPort = resources.Pulp.Spec.Database.PostgresPort
-		}
-		dbHost = resources.Pulp.Name + "-database-svc"
-		dbPort = strconv.Itoa(containerPort)
-
-		postgresEnvVars := []corev1.EnvVar{
-			{Name: "POSTGRES_SERVICE_HOST", Value: dbHost},
-			{Name: "POSTGRES_SERVICE_PORT", Value: dbPort},
-		}
-		envVars = append(envVars, postgresEnvVars...)
-	} else {
-		postgresEnvVars := []corev1.EnvVar{
-			{
-				Name: "POSTGRES_SERVICE_HOST",
-				ValueFrom: &corev1.EnvVarSource{
-					SecretKeyRef: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: resources.Pulp.Spec.Database.ExternalDBSecret,
-						},
-						Key: "POSTGRES_HOST",
-					},
-				},
-			}, {
-				Name: "POSTGRES_SERVICE_PORT",
-				ValueFrom: &corev1.EnvVarSource{
-					SecretKeyRef: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: resources.Pulp.Spec.Database.ExternalDBSecret,
-						},
-						Key: "POSTGRES_PORT",
-					},
-				},
-			},
-		}
-		envVars = append(envVars, postgresEnvVars...)
-	}
-
-	// add cache configuration if enabled
-	if resources.Pulp.Spec.Cache.Enabled {
-
-		// if there is no ExternalCacheSecret defined, we should
-		// use the redis instance provided by the operator
-		if len(resources.Pulp.Spec.Cache.ExternalCacheSecret) == 0 {
-			var cacheHost, cachePort string
-
-			if resources.Pulp.Spec.Cache.RedisPort == 0 {
-				cachePort = strconv.Itoa(6379)
-			} else {
-				cachePort = strconv.Itoa(resources.Pulp.Spec.Cache.RedisPort)
-			}
-			cacheHost = resources.Pulp.Name + "-redis-svc." + resources.Pulp.Namespace
-
-			redisEnvVars := []corev1.EnvVar{
-				{Name: "REDIS_SERVICE_HOST", Value: cacheHost},
-				{Name: "REDIS_SERVICE_PORT", Value: cachePort},
-			}
-			envVars = append(envVars, redisEnvVars...)
-		} else {
-			redisEnvVars := []corev1.EnvVar{
-				{
-					Name: "REDIS_SERVICE_HOST",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: resources.Pulp.Spec.Cache.ExternalCacheSecret,
-							},
-							Key: "REDIS_HOST",
-						},
-					},
-				}, {
-					Name: "REDIS_SERVICE_PORT",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: resources.Pulp.Spec.Cache.ExternalCacheSecret,
-							},
-							Key: "REDIS_PORT",
-						},
-					},
-				}, {
-					Name: "REDIS_SERVICE_DB",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: resources.Pulp.Spec.Cache.ExternalCacheSecret,
-							},
-							Key: "REDIS_DB",
-						},
-					},
-				}, {
-					Name: "REDIS_SERVICE_PASSWORD",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: resources.Pulp.Spec.Cache.ExternalCacheSecret,
-							},
-							Key: "REDIS_PASSWORD",
-						},
-					},
-				},
-			}
-			envVars = append(envVars, redisEnvVars...)
-		}
-	}
-
-	if resources.Pulp.Spec.SigningSecret != "" {
-
-		// for now, we are just dumping the error, but we should handle it
-		signingKeyFingerprint, _ := resources.RepoManagerReconciler.getSigningKeyFingerprint(resources.Pulp.Spec.SigningSecret, resources.Pulp.Namespace)
-
-		signingKeyEnvVars := []corev1.EnvVar{
-			{Name: "PULP_SIGNING_KEY_FINGERPRINT", Value: signingKeyFingerprint},
-			{Name: "COLLECTION_SIGNING_SERVICE", Value: getPulpSetting(resources.Pulp, "galaxy_collection_signing_service")},
-			{Name: "CONTAINER_SIGNING_SERVICE", Value: getPulpSetting(resources.Pulp, "galaxy_container_signing_service")},
-		}
-		envVars = append(envVars, signingKeyEnvVars...)
-	}
-
-	dbFieldsEncryptionSecret := ""
-	if resources.Pulp.Spec.DBFieldsEncryptionSecret == "" {
-		dbFieldsEncryptionSecret = resources.Pulp.Name + "-db-fields-encryption"
-	} else {
-		dbFieldsEncryptionSecret = resources.Pulp.Spec.DBFieldsEncryptionSecret
-	}
-
-	adminSecretName := resources.Pulp.Name + "-admin-password"
-	if len(resources.Pulp.Spec.AdminPasswordSecret) > 1 {
-		adminSecretName = resources.Pulp.Spec.AdminPasswordSecret
-	}
-
-	volumes := []corev1.Volume{
-		{
-			Name: resources.Pulp.Name + "-server",
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName: resources.Pulp.Name + "-server",
-					Items: []corev1.KeyToPath{{
-						Key:  "settings.py",
-						Path: "settings.py",
-					}},
-				},
-			},
-		},
-		{
-			Name: adminSecretName,
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName: adminSecretName,
-					Items: []corev1.KeyToPath{{
-						Path: "admin-password",
-						Key:  "password",
-					}},
-				},
-			},
-		},
-		{
-			Name: resources.Pulp.Name + "-db-fields-encryption",
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName: dbFieldsEncryptionSecret,
-					Items: []corev1.KeyToPath{{
-						Key:  "database_fields.symmetric.key",
-						Path: "database_fields.symmetric.key",
-					}},
-				},
-			},
-		},
-	}
-
-	_, storageType := controllers.MultiStorageConfigured(resources.Pulp, "Pulp")
-
-	// if SC defined, we should use the PVC provisioned by the operator
-	if storageType[0] == controllers.SCNameType {
-		fileStorage := corev1.Volume{
-			Name: "file-storage",
-			VolumeSource: corev1.VolumeSource{
-				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-					ClaimName: resources.Pulp.Name + "-file-storage",
-				},
-			},
-		}
-		volumes = append(volumes, fileStorage)
-
-		// if .spec.Api.PVC defined we should use the PVC provisioned by user
-	} else if storageType[0] == controllers.PVCType {
-		fileStorage := corev1.Volume{
-			Name: "file-storage",
-			VolumeSource: corev1.VolumeSource{
-				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-					ClaimName: resources.Pulp.Spec.PVC,
-				},
-			},
-		}
-		volumes = append(volumes, fileStorage)
-
-		// if there is no SC nor PVC nor object storage defined we will mount an emptyDir
-	} else if storageType[0] == controllers.EmptyDirType {
-		emptyDir := []corev1.Volume{
-			{
-				Name: "tmp-file-storage",
-				VolumeSource: corev1.VolumeSource{
-					EmptyDir: &corev1.EmptyDirVolumeSource{},
-				},
-			},
-			{
-				Name: "assets-file-storage",
-				VolumeSource: corev1.VolumeSource{
-					EmptyDir: &corev1.EmptyDirVolumeSource{},
-				},
-			},
-		}
-		volumes = append(volumes, emptyDir...)
-	}
-
-	if resources.Pulp.Spec.SigningSecret != "" {
-		signingSecretVolume := []corev1.Volume{
-			{
-				Name: resources.Pulp.Name + "-signing-scripts",
-				VolumeSource: corev1.VolumeSource{
-					ConfigMap: &corev1.ConfigMapVolumeSource{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: resources.Pulp.Spec.SigningScriptsConfigmap,
-						},
-					},
-				},
-			},
-			{
-				Name: resources.Pulp.Name + "-signing-galaxy",
-				VolumeSource: corev1.VolumeSource{
-					Secret: &corev1.SecretVolumeSource{
-						SecretName: resources.Pulp.Spec.SigningSecret,
-						Items: []corev1.KeyToPath{
-							{
-								Key:  "signing_service.gpg",
-								Path: "signing_serivce.gpg",
-							},
-							{
-								Key:  "signing_service.asc",
-								Path: "signing_serivce.asc",
-							},
-						},
-					},
-				},
-			},
-		}
-		volumes = append(volumes, signingSecretVolume...)
-	}
-
-	var containerAuthSecretName string
-	if resources.Pulp.Spec.ContainerTokenSecret != "" {
-		containerAuthSecretName = resources.Pulp.Spec.ContainerTokenSecret
-	} else {
-		containerAuthSecretName = resources.Pulp.Name + "-container-auth"
-	}
-
-	containerTokenSecretVolume := corev1.Volume{
-		Name: resources.Pulp.Name + "-container-auth-certs",
-		VolumeSource: corev1.VolumeSource{
-			Secret: &corev1.SecretVolumeSource{
-				SecretName: containerAuthSecretName,
-				Items: []corev1.KeyToPath{
-					{
-						Key:  "container_auth_public_key.pem",
-						Path: resources.Pulp.Spec.ContainerAuthPublicKey,
-					},
-					{
-						Key:  "container_auth_private_key.pem",
-						Path: resources.Pulp.Spec.ContainerAuthPrivateKey,
-					},
-				},
-			},
-		},
-	}
-	volumes = append(volumes, containerTokenSecretVolume)
-
-	volumeMounts := []corev1.VolumeMount{
-		{
-			Name:      resources.Pulp.Name + "-server",
-			MountPath: "/etc/pulp/settings.py",
-			SubPath:   "settings.py",
-			ReadOnly:  true,
-		},
-		{
-			Name:      adminSecretName,
-			MountPath: "/etc/pulp/pulp-admin-password",
-			SubPath:   "admin-password",
-			ReadOnly:  true,
-		},
-		{
-			Name:      resources.Pulp.Name + "-db-fields-encryption",
-			MountPath: "/etc/pulp/keys/database_fields.symmetric.key",
-			SubPath:   "database_fields.symmetric.key",
-			ReadOnly:  true,
-		},
-	}
-
-	// we will mount file-storage if a storageclass or a pvc was provided
-	if storageType[0] == controllers.SCNameType || storageType[0] == controllers.PVCType {
-		fileStorageMount := corev1.VolumeMount{
-			Name:      "file-storage",
-			ReadOnly:  false,
-			MountPath: "/var/lib/pulp",
-		}
-		volumeMounts = append(volumeMounts, fileStorageMount)
-
-		// if no file-storage nor object storage were provided we will mount the emptyDir
-	} else if storageType[0] == controllers.EmptyDirType {
-		emptyDir := []corev1.VolumeMount{
-			{
-				Name:      "tmp-file-storage",
-				MountPath: "/var/lib/pulp/tmp",
-			},
-			{
-				Name:      "assets-file-storage",
-				MountPath: "/var/lib/pulp/assets",
-			},
-		}
-		volumeMounts = append(volumeMounts, emptyDir...)
-	}
-
-	if resources.Pulp.Spec.SigningSecret != "" {
-		signingSecretMount := []corev1.VolumeMount{
-			{
-				Name:      resources.Pulp.Name + "-signing-scripts",
-				MountPath: "/var/lib/pulp/scripts",
-				SubPath:   "scripts",
-				ReadOnly:  true,
-			},
-			{
-				Name:      resources.Pulp.Name + "-signing-galaxy",
-				MountPath: "/etc/pulp/keys/signing_service.gpg",
-				SubPath:   "signing_service.gpg",
-				ReadOnly:  true,
-			},
-			{
-				Name:      resources.Pulp.Name + "-signing-galaxy",
-				MountPath: "/etc/pulp/keys/singing_service.asc",
-				SubPath:   "signing_service.asc",
-				ReadOnly:  true,
-			},
-		}
-		volumeMounts = append(volumeMounts, signingSecretMount...)
-	}
-
-	if resources.Pulp.Spec.ContainerTokenSecret != "" {
-		containerTokenSecretMount := []corev1.VolumeMount{
-			{
-				Name:      resources.Pulp.Name + "-container-auth-certs",
-				MountPath: "/etc/pulp/keys/container_auth_private_key.pem",
-				SubPath:   "container_auth_private_key.pem",
-				ReadOnly:  true,
-			},
-			{
-				Name:      resources.Pulp.Name + "-container-auth-certs",
-				MountPath: "/etc/pulp/keys/container_auth_public_key.pem",
-				SubPath:   "container_auth_public_key.pem",
-				ReadOnly:  true,
-			},
-		}
-		volumeMounts = append(volumeMounts, containerTokenSecretMount...)
-	}
-
-	// mountCASpec adds the trusted-ca bundle into []volume and []volumeMount if pulp.Spec.TrustedCA is true
-	if IsOpenShift {
-		volumes, volumeMounts = mountCASpec(resources.Pulp, volumes, volumeMounts)
-	}
-
-	resourceRequirements := resources.Pulp.Spec.Api.ResourceRequirements
-
-	readinessProbe := resources.Pulp.Spec.Api.ReadinessProbe
-	if readinessProbe == nil {
-		readinessProbe = &corev1.Probe{
-			ProbeHandler: corev1.ProbeHandler{
-				Exec: &corev1.ExecAction{
-					Command: []string{
-						"/usr/bin/readyz.py",
-						getPulpSetting(resources.Pulp, "api_root") + "api/v3/status/",
-					},
-				},
-			},
-			FailureThreshold:    10,
-			InitialDelaySeconds: 60,
-			PeriodSeconds:       10,
-			SuccessThreshold:    1,
-			TimeoutSeconds:      10,
-		}
-	}
-
-	livenessProbe := resources.Pulp.Spec.Api.LivenessProbe
-	if livenessProbe == nil {
-		livenessProbe = &corev1.Probe{
-			FailureThreshold: 10,
-			ProbeHandler: corev1.ProbeHandler{
-				HTTPGet: &corev1.HTTPGetAction{
-					Path: getPulpSetting(resources.Pulp, "api_root") + "api/v3/status/",
-					Port: intstr.IntOrString{
-						IntVal: 24817,
-					},
-					Scheme: corev1.URIScheme("HTTP"),
-				},
-			},
-			InitialDelaySeconds: 120,
-			PeriodSeconds:       20,
-			SuccessThreshold:    1,
-			TimeoutSeconds:      10,
-		}
-	}
-
-	// the following variables are defined to avoid issues with reconciliation
-	restartPolicy := corev1.RestartPolicy("Always")
-	terminationGracePeriodSeconds := int64(30)
-	dnsPolicy := corev1.DNSPolicy("ClusterFirst")
-	schedulerName := corev1.DefaultSchedulerName
-	Image := os.Getenv("RELATED_IMAGE_PULP")
-	if len(resources.Pulp.Spec.Image) > 0 && len(resources.Pulp.Spec.ImageVersion) > 0 {
-		Image = resources.Pulp.Spec.Image + ":" + resources.Pulp.Spec.ImageVersion
-	} else if Image == "" {
-		Image = "quay.io/pulp/pulp-minimal:stable"
-	}
-
-	// deployment definition
-	dep := &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      resources.Pulp.Name + "-api",
-			Namespace: resources.Pulp.Namespace,
-			Annotations: map[string]string{
-				"email": "pulp-dev@redhat.com",
-				"ignore-check.kube-linter.io/no-node-affinity": "Do not check node affinity",
-			},
-			Labels: map[string]string{
-				"app.kubernetes.io/name":       resources.Pulp.Spec.DeploymentType + "-api",
-				"app.kubernetes.io/instance":   resources.Pulp.Spec.DeploymentType + "-api-" + resources.Pulp.Name,
-				"app.kubernetes.io/component":  "api",
-				"app.kubernetes.io/part-of":    resources.Pulp.Spec.DeploymentType,
-				"app.kubernetes.io/managed-by": resources.Pulp.Spec.DeploymentType + "-operator",
-				"app":                          "pulp-api",
-				"pulp_cr":                      resources.Pulp.Name,
-				"owner":                        "pulp-dev",
-			},
-		},
-		Spec: appsv1.DeploymentSpec{
-			Replicas: &replicas,
-			Strategy: strategy,
-			Selector: &metav1.LabelSelector{
-				MatchLabels: ls,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: ls,
-				},
-				Spec: corev1.PodSpec{
-					Affinity:                  affinity,
-					SecurityContext:           podSecurityContext,
-					NodeSelector:              nodeSelector,
-					Tolerations:               toleration,
-					Volumes:                   volumes,
-					ServiceAccountName:        resources.Pulp.Name,
-					TopologySpreadConstraints: topologySpreadConstraint,
-					Containers: []corev1.Container{{
-						Name:            "api",
-						Image:           Image,
-						ImagePullPolicy: corev1.PullPolicy(resources.Pulp.Spec.ImagePullPolicy),
-						Args:            []string{"pulp-api"},
-						Env:             envVars,
-						Ports: []corev1.ContainerPort{{
-							ContainerPort: 24817,
-							Protocol:      "TCP",
-						}},
-						LivenessProbe:  livenessProbe,
-						ReadinessProbe: readinessProbe,
-						Resources:      resourceRequirements,
-						VolumeMounts:   volumeMounts,
-					}},
-
-					/* the following configs are not defined on pulp-operator (ansible version)  */
-					/* but i'll keep it here just in case we can manage to make deepequal usable */
-					RestartPolicy:                 restartPolicy,
-					TerminationGracePeriodSeconds: &terminationGracePeriodSeconds,
-					DNSPolicy:                     dnsPolicy,
-					SchedulerName:                 schedulerName,
-				},
-			},
-		},
-	}
-
-	// Set Pulp instance as the owner and controller
-	ctrl.SetControllerReference(resources.Pulp, dep, resources.RepoManagerReconciler.Scheme)
-	return dep
 }
 
 // labelsForPulpApi returns the labels for selecting the resources
@@ -979,7 +387,7 @@ MEDIA_ROOT = ""
 	// configure TOKEN_SERVER based on ingress_type
 	tokenServer := "http://" + resources.Pulp.Name + "-api-svc." + resources.Pulp.Namespace + ".svc.cluster.local:24817/token/"
 	if resources.Pulp.Spec.IngressType == "route" {
-		tokenServer = "https://" + getRouteHost(resources) + "/token/"
+		tokenServer = rootUrl + "/token/"
 	} else if resources.Pulp.Spec.IngressType == "ingress" {
 		proto := "http"
 		if len(resources.Pulp.Spec.IngressTLSSecret) > 0 {
